Move HTTP client coroutine API to base class
