+++ lib/inc/drogon/HttpClient.h
@@ -23,6 +23,7 @@
 #include <functional>
 #include <memory>
 #include <future>
+#include "drogon/HttpBinder.h"
 
 #ifdef __cpp_impl_coroutine
 #include <drogon/utils/coroutine.h>
@@ -32,6 +33,27 @@ namespace drogon
 {
 class HttpClient;
 using HttpClientPtr = std::shared_ptr<HttpClient>;
+
+namespace internal
+{
+#ifdef __cpp_impl_coroutine
+struct HttpRespAwaiter : public CallbackAwaiter<HttpResponsePtr>
+{
+    HttpRespAwaiter(HttpClient *client, HttpRequestPtr req, double timeout)
+        : client_(client), req_(std::move(req)), timeout_(timeout)
+    {
+    }
+
+    void await_suspend(std::coroutine_handle<> handle);
+
+  private:
+    HttpClient *client_;
+    HttpRequestPtr req_;
+    double timeout_;
+};
+#endif
+
+}  // namespace internal
 
 /// Asynchronous http client
 /**
@@ -131,8 +153,13 @@ class HttpClient : public trantor::NonCopyable
      *
      * @return task<HttpResponsePtr>
      */
-    virtual Task<HttpResponsePtr> sendRequestCoro(HttpRequestPtr req,
-                                                  double timeout = 0) = 0;
+    Task<HttpResponsePtr> sendRequestCoro(HttpRequestPtr req,
+                                          double timeout = 0)
+    {
+        co_return co_await internal::HttpRespAwaiter(this,
+                                                     std::move(req),
+                                                     timeout);
+    }
 #endif
 
     /// Set the pipelining depth, which is the number of requests that are not
@@ -236,7 +263,42 @@ class HttpClient : public trantor::NonCopyable
     }
 
   protected:
+    void await_suspend(std::coroutine_handle<> handle);
     HttpClient() = default;
 };
+
+#ifdef __cpp_impl_coroutine
+inline void internal::HttpRespAwaiter::await_suspend(
+    std::coroutine_handle<> handle)
+{
+    client_->sendRequest(
+        req_,
+        [handle = std::move(handle), this](ReqResult result,
+                                           const HttpResponsePtr &resp) {
+            if (result == ReqResult::Ok)
+            {
+                setValue(resp);
+                handle.resume();
+            }
+            else
+            {
+                std::string reason;
+                if (result == ReqResult::BadResponse)
+                    reason = "BadResponse";
+                else if (result == ReqResult::NetworkFailure)
+                    reason = "NetworkFailure";
+                else if (result == ReqResult::BadServerAddress)
+                    reason = "BadServerAddress";
+                else if (result == ReqResult::Timeout)
+                    ;
+                reason = "Timeout";
+                setException(
+                    std::make_exception_ptr(std::runtime_error(reason)));
+                handle.resume();
+            }
+        },
+        timeout_);
+}
+#endif
 
 }  // namespace drogon
